修改了结构，声明和实现基本分开，增加了kafka的Appender
